refactor: better helpers organization
